CORE-7637: moved the implementation of the resource_types endpoints in the permissions service to a separate package
