+++ golang/src/permissions/restapi/impl/resource_types/put_resource_types.go
@@ -1,4 +1,4 @@
-package impl
+package resource_types
 
 import (
 	"database/sql"
